Rename DebugServiceRegistration to DebugServiceProviders
